requirements.txt
psycopg2-binary==2.9.9
pgvector==0.2.3
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3 